requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
jinja2==3.1.6
python-multipart==0.0.20
sqlmodel==0.0.22
mistune==3.1.3
bleach==6.2.0
pydantic-settings==2.8.1
